backend/requirements.txt
fastapi
pydantic
chromadb
groq
openai
tiktoken
uvicorn
langchain-openai
